Extract WidgetSelect to its own file
